pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "talent-graph"
version = "0.1.0"
description = "Talent Discovery platform using knowledge graphs, embeddings, and LLM-powered explanations"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",
    "neo4j>=5.27.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "python-ulid>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "respx>=0.21.0",
    "factory-boy>=3.3.1",
    "testcontainers[postgres,neo4j]>=4.9.0",
]

[project.scripts]
talent-graph = "talent_graph.scripts.seed_ingest:main"

[tool.hatch.build.targets.wheel]
packages = ["src/talent_graph"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["--cov=src/talent_graph", "--cov-report=term-missing", "-v"]

[tool.coverage.run]
source = ["src/talent_graph"]
omit = ["*/tests/*", "*/migrations/*"]
